Using templates
I created the templates directory and then the blog subdirectory. I also made HTML in order to create the about and
the home blog page using a little bit of the Django Templates Language. I iterated through dummy data in the context
(The way I use to pass information and Python objects in HTML) and displayed the context data on screen