Creating the User Registration Form
Given that creating a user registration form is a frecuent task, Django provides a class that make that form for me, so
I had to import the UserCreationForm class and cosntruct an object from that blueprint within the view. Then, I render
the template passing the form object in the context.

The url pattern that I made was 'register/' and the view that handles this pattern is register from users app. This time,
I wanted to set up the urlpattern in the main urls.py file, so I imported the views.py from users folder as users_views
because I could made more patterns in that file in the future so naming views.py that way allows me to be relaxed and
have code flexibility