Redirecting the Post Request and Adding a Flash Message
Given that I didn't specify an url in the action form attribute, the form posts back the data in the same route, so
I had to put an if statement that handles when the request method is post and when the method is get. I can access the
method with request.method. If that happens, I want to create a form with the request.POST data, then validate the form
data with the is_valid() method.

Once the form is validated, I grab the userName from the cleaned_data form dictionary key called 'username' and then
I created a success flash message. Finally, I redirect the user to the blog url. The messages are sent without putting
them in any context.

In the html base template (because I want to put flash messages at all the web pages) I added an if statement that
checks the messages, then loop through the messages and put every single message in a div tag that has some Bootstrap
classes (alert and alert-type of flash message). Since the flash message type in Django is the same as Bootstrap, I
passed the message tags attribute in the class