Creating the profile view and template
- MODEL:
This model has a OneToOne field with the User Model and it also has an ImageField with a default image and the location
where I want to upload the file. Then, I run the migrations and I added the model to the admin site

- MEDIA_ROOT and MEDIA_URL in settings.py:
I had to add these variables in the settings.py file in order to define what the media folder is and how the browser
is gonna look for the image source

- HTML File:
This file displays on screen the username, email and the user profile picture. I can access the picture source through
the user object like this: 'user.profile.image.url'
 #	new file:   django_project/media/default.jpg