Creating the search bar and the default profile creation
The searchbar HTML was simple. This only has an input text field and a button. The form doesn't have the method attribute
so, the data will be sent in the Http GET method appended at the end of the url. The view will get the search value
from the form, otherwise, it will be false. If there is the search attribute, Django does a queryset and it also
gets a list of post objects that will pass into the context.

The default profile creation is made with Django signals. This is the step by step:
1. Create a signals.py file in the app folder
2. Import the signal, sender, receiver and Profile model
3. Create the receiver function and pass the required arguments. In this case, the arguments are sender, instance, created
and **kwargs
4. Add the receiver decorator above the receiver function. This decorator uses the signal as the first argument and
the sender as the second argument
5. Create the ready method in apps.py
6. Import there the signals