Creating the user update feature
In order to make this feature great, I had to create the forms, edit the view and put the form HTML tags

1. Forms.py:

Given that I want to update the profile image and the user information, I had to create two model forms cuz the project
has the profile and user in different models. In the Meta class I specified which Model I want to work with and the
form fields in the 'fields' list

2. Views.py
The view handles the POST and GET Http Requests. If the request method is post, the view creates the two model forms
and it fills them with the POST data and FILES data (in the case of the profile form). It also specifies the
instance parameter in the moment where the Form is created because I want to set the user and profile objects in order
to update them in the database if the form is valid.

If the request is 'GET', the view creates the forms and fills them with the current data in the database sstting the
instance parameter.