Create, Detail and Update Views
1. Create:
I used the generic editing views provided by Django, so I created the PostCreateView that inherits from the Django
Create View. I also used the Login Required Mixin in order to give access just to the users that are logged in.
Given that the form displayed on screen only has the title and content fields, I had to override the form_valid method
to add the request user in the form.instance object.

The View passes a form object into the template. This template puts the form with crispy tags

2. Update:
The update view is about the same as the Create view, but this time, I had to override the get_queryset method
in order to filter the super queryset by the post author, because I want the post author to be the same request user.
If that condition is False, Django sends HTTP Error

3. Detail:
The Detail View gets the object with the primary key passed in the url and puts the post into the detail template.
This template shows the post information in a bootstrap card. If the user is the same post author, it also
shows the edit and delete button, given that I only want the post author to have that permission.