Updating the Profile View
I wanted to allow some user to watch another user profile, because, before I just had the profile view for the current
logged in user. This time I updated the user form and I added the first_name and last_name fields.

The view looks for the user object by its id passed in the url and takes that object to the HTML template that
shows on screen the profile info and then checks whether or not the user object is the same logged in user.
If that is true, the user can see the User and Profile form, otherwise, the template doesn't show the forms.