Delete View and CSS styling
I wanted to preserve the whitespace when the user saves something in the database, so I used CSS to display the text
on screen with thid feature. The class .article-content is selected in 'mainn.css' and the style rule is:
'white-space: pre-line;'

The DeleteView is a Django Generic View and this view looks for the object by its primary key and then passes it to
the delete.html template that asks the users if they really want to delete the post.