Adding the favorites feature
This feature allows the users to mark posts as favorites and save it in the database. Then, if the users want to look
for their favorites posts, they just need to click on the favorites link on the navbar

- MODEL: The basic idea is create a new model that is a middle tab between the User and the Post Model. This Favorites
tab is going to handle the Many To Many relationship, given that a user can have multiple posts as favorites and
a post can be a favorite of many users. The Posts model only needs another field and that is the favs field with
the models.ManyToManyField(User, through="Favorite", related_name="all_favs"). The through parameter specifies
what the middle table is and the related_name allows me to get a queryset object with all the posts that are favorites
of some user.

The Favorite Model needs the Foreign Keys of the user and post model. This also has the Meta class in order to add
some settings to the class. This time, 'unique_together' is a constraint that doesn't allow to store the same user
and post combination. It doesn't make sense to store that a user saves some post as favorite two times

-VIEWS:
There are two views that handles whether or not a user 'favs' some post. If the user favs a post, the view gets
an post object based on the primary key passed in the url, creating a new Favorite object and storing it in the
database. If the users wanted to 'unfav' some post, the view also gets a post object, the favorite object
based on the post and the user that is making the request and delete it from the database. Both views return an
HttpResponse object, given that Django expects a response object for each view.

There is also a favorites view that looks for all the user favs and it passes the list of posts to the template through
the context

-TEMPLATES:

The template checks if the post that is going to be displayed on screen it's in the users favorites list. If that is
Ture, it puts an anchor tag with the favorite icon with a 'btn-favorite' class, else, it puts the 'btn-no-favorite'
class in the HTML tag. Django also gives the tag an id according to the post primary key, and creates the 'onclick'
attribute that calls the JS function 'handleFavorites' passing to it the urls to fav or un fav the post and the
post id