Creating the delete comment feature more friendly
In order to create that feature more attractive, I created an onclick event in the delete anchor tag that passes
the delete url, the comment id and a csrf token to the deleteComment JavaScript Function. The url is used to
create a post request and send it through the XMLHttpRequest AJAX object, the comment id is used to get the li tag, then,
I deleted it from the Document Object Model with the removeChild method.

Finally, the csrf token is sent through the request header called 'X-CSRFToken' that was added in the xhttp object through
the setRequestHeader method