Creating the chat room application
There are two main parts of this feature:

1. TEMPLATE: It has a script tag that connects the client with the WebSocket Server created with Django Channels. The socket object waits for
the server's message and then it parses the JSON object and displays the username and its message on screen. There is an eent listener
waiting for a click in the button, so that the browser gets the message and send it with the send method to the server.

2. CONSUMERS: In the other hand, consumers.py is used to set the chat room from the url (in the scope dictionary), then, it sets the group
with the group_add method. The main part of this is the receive method that gets the JSON object from the client, parses it and then
it generates an event sending the information to another method called "chat_message". That way, the information is broadcasted to
all the users that are in a certain group.