Notifications System
In order to create a notification system that gets triggered when a user follows somebody, comments a post or sends a message, I had to
create a web sockets connection in the first page the user loads (that's base.html). The connection is stablished in when the corresponding
connect method in the UserConsumer accepts it, and then, this consumer creates a group with the user id, so that only the user is in
that group and receives notifications when the user broadcasts an event.

When some user triggers the new_follow event, that method catches the signal and then it sends a JSON object with the notification data
to the websocket client. The client receives that in the onmessage method, it reads the data with JSON.parse and creates a tag that
contains the notification info and push it into the template.

I also created a new template tag that accepts the context, gets all the user notifications and sent that to the template that is going
to be included. This template put the information in a dropdown html element. Finally, this dropdown is included in all the pages.