Handling New Notifications
In 'blog/signals.py' I added a new signal that gets triggered when a user comments a post. This signal creates a new notification object
if the user is making a new comment in the post of another user. Then, the post_save notification signal also triggers another function
that sends a message to the channels consumer. I did the same process with 'chat/signals.py' when some user sends a message.

The other change was using the get_url method on the instance in 'notifications/signals.py' because it was hardcoded. That method
gets the desired url for ant notification it receives.