Notification View
I created a new view that gets all the user notifications and send them to the template. It also has a paginator in
order to send different pages by twenty notifications.

The template iterates through each notification and display the corresponding information. This template has
JavaScript in order to use an infinite scroll system, so the page gets the next page notifications without
refreshing the page. This functionality is added with the Waypoint object. The code is implemented acording with
the waypoint documentation.