Mark notification as seen
The main idea is that the notification is marked as seen when the user goes for the event page. Each notification in the notifications
list panel has an anchor tag that sends the user to the mark_notification view. This view changes the state in the database and then redirects
to the event page, whether or not is the comment, chat or profile page (according to the notification).

I also had to change the signals because I want to put the new url view in the notification item that is created by the websocket client
when receives an event.