Adding search functionality
I added search functionality to the profile and notifications list view, so that the user can get the notifications sent by some specific user
and the user also can get the user profile by its username or email in the searchbar.

In order to do this, I had to use the Q filter, passing the field (username and email) and the lookup query that I wanted to do, in this case,
icontains. Then, I pass that query to the filter method used in the context manager. This search functionality is also handled by the view
getting the search key from the url if this exists with request.GET.get("search", False).