Blog Views Updated
The home page view has new functionality such as pagination. This is implemented with the paginator class provided by Django, that way,
I can group data and send faster web pages. This view also has a query when users send data by the search form located in the navbar,
so that they can filter posts by their content. Finally, this view sends a list of users that my current logged in user doesn't follow
as suggestions.

The post create view is used to renders a form and then it saves that data. The form_valid method is used to add the user author in
the post object, so that this information is saved in the database. Then, the post update view and delete view uses the same
method (get_queryset) in order to filter the data by the author. If the author is not the user making that request, that user
is gonna receive an error. I did this in order to provide security, because, a user can go to a specific hardcoded url and this could
be dangerous.