Extra Functionality
I added some fatures to the post detail page: the first one was putting the comment options in a dropdown, so the comment owner can
view the options (delete and edit) grouped by a Bootstrap Dropdown. The other feature was adding some paginations to the comments.
The view creates a paginator object and group the comments by five, depending on the page url argument. In the template I use
waypoints, adding a class "infinite-container" to the div that contains each commment (this also has the class "infinite-item".
Finally, I created a waypoint.Infinite object specifying the list container, each time the user goes to the bottom, it automatically
gets the comments of the next page and displays them on screen.

The other feature was updating the follow and unFollow functions in "scripts.js", so each time the user follows a profile, the function creates
the message button and insert it in the template. In the other hand, the unFollow method gets the message button and removes it from the template
with the remove method.