Like functionality
MODELS: The main idea was creating a many to many relationship, given that one user can like many posts and one post can be liked by
many users. The model has the constraint unique_together, indicating that the user and post saved in the database row has to be unique.

VIEWS: I added two views to handle the like and unlike process. When a user sends a post request to the like url pattern, the LikeView
gets the post from the url and saves the new like in the database. In the other hand, the UnLikeView looks for the Like saved
and deletes it if exists.

TEMPLATE: The template has a new star icon that uses the handleFavorite function each time the user clicks on it. This function sends
the post request to the server and updates the corresponding CSS class for like or unlike.