Like Notification
In order to add a like notification, I had to add the post field in the Notification model, given that I want to have the post liked by the user.
Then, I had to add a signal when a Like is created. This checks if the user is the same as the post author, because I don't want to create a
Notification when the users are liking their own post. Then, it checks if a notification field was created before by the same user receiving
the like in the same post. If that's the case, I don't want to create another like notification for the same post, I just want to update
the notification adding a new text preview indicating the last user and the total number of likes this post has. It also changes the
has_seen attribute to False and save the notification.

If the notification doesn't exists, the signal creates a new one, specifying the sender, the receiver, notification type as four (for likes),
the post liked and the text_preview used to show a coherent message to the user. Finally, in signals.py from the notifications app I removed the
if created clause, because this checks if a new instance is created, but this doesn't handle whenever a notification field is updated,
and I also want to show a real time notification when the object is updated.