Trending feature
The trending feature shows the most engaging posts created by users in a short period of time. The view gets a queryset filtering
by date and then it selects the posts with a lot of comments and likes and order them using the order_by clause.

TEMPLATE: The template receives the list of the five more popular posts and then displays the author, likes, comments and the title
in an anchor tag, so the user can go to that post and get hooked by the content.