Chat App
I added chat pagination for the messages. When the user goes to some chat page, only the last ten messages ar displayed on screen. This is made
with the Django paginator, but the page that is sent is the last one, because I want to render the last ten messages, not the first ones. The
corresponding page is found with the expression in the line 94 "page_number + number_of_pages + 1". Let's say we have five pages, the first page
the user wants to see is the last one, the second page is the fourth one and so on, that expression gives me the desired page no matter how
many pages the chat is divided in.

In scripts.js located in 'chat' folder, it was created a waypoint to get signals every time the user goes up in the chat, so the frontend
triggers a new fetch request in order to receive the next page messages, then it puts the response in the msg_container using the
insertAdjacentHTML method on that container. All the element information is passed with outerHTML because this attribute gets the element
as a string