AWS File Storage System
In settings.py I had to add a new storage app in INSTALLED_APPS that is going to handle the project images and save them in my S3 bucket. I also
added some variables that the django-storage app uses, getting their values from environment variables.

This changes led me to replace some functionality, for example, the new backend storage doesn't use the image.path attribute, so I had to replace
it and work with image.name. In signals.py I had to use storage.delete(image.name) instead of os.remove, because I'm not removing images from
my OS anymore. The sae method from the Profile model was replaced for the save method of the ProfileUpdateForm. The cropped image has to be saved
in a BytesIO object, and then, this information can be placed in S3 with storage.save method.