Deployment Configuration
-settings.py:
This file needs some changes for deployment: CHANNELS_LAYERS variable was added according to the channels documentation, it says the backend
is channels_redis, so I needed to do a pip install channels-redis and put it in requirements.txt. I also put the domain 'channelsapp.herokuapp'
in ALLOWED_HOSTS in order to give that host the permissions for running the application. Finally, I added the STATIC_ROOT that
Heroku is going to use when it runs the 'collectstatic' command.

-PROCFILE: Heroku needs a Procfile to kwno how I want to run my application. I specified the asgi app and daphne as backend server.

-runtime: This txt file is used to specify the Python version